move logic into each get*Filter function to improve readability
